DZExplorer_Backend/requirements.txt
<<<<<<< HEAD
asgiref==3.7.1
bcrypt==4.0.1
dj-database-url==2.0.0
Django==4.2.1
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
whitenoise==6.4.0
=======
Django
>>>>>>> f58f656431231fa001d132305f1903d29c2de228
